nsqtest/docker: fix bug related with timeout on initial docker container deploy
Fixed a bug which caused flaky tests related with nsq docker container deployment. Every time the test machine was restarted or docker engine stopped, the first time test are run, it times out when trying to connect to the container. This happened ONLY on the first tests run after restart.